+++ nemo/collections/common/parts/utils.py
@@ -96,3 +96,12 @@ def flatten(list_in: List) -> List:
         A flat list of values.
     """
     return list(flatten_iterable(list_in))
+
+
+def extend_instance(obj, mixin):
+    """Apply mixins to a class instance after creation"""
+    base_cls = obj.__class__
+    base_cls_name = obj.__class__.__name__
+    obj.__class__ = type(
+        base_cls_name, (mixin, base_cls), {}
+    )  # mixin needs to go first for our forward() logic to work
